refactor: change structure of application
