+++ src/components/layout.css
@@ -17,10 +17,6 @@ h1,h2,h3,h4,h5,.body-text {
   margin-right: auto;
   width: 300px;
 }
-.map-container {
-  display:  flex;
-  justify-content: center;
-}
 article,
 aside,
 details,
