+++ src/components/map/map.css
@@ -1,0 +1,4 @@
+.map-container {
+    display:  flex;
+    justify-content: center;
+  }
