+++ src/components/video/video.js
@@ -2,8 +2,8 @@ import PropTypes from "prop-types";
 import React, { useState } from "react";
 import { I18n } from 'react-i18next';
 import YouTube from 'react-youtube';
-import ModalComponent from './modal';
-import Button from './button';
+import ModalComponent from '../modal/modal';
+import Button from '../button/button';
 import './video.css';
 
 const Video = ({ videoId, children }) => {
